# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Proof Fabric Protocol (PFP) API
  description: Cryptographically verifiable Financial Evidence Artifacts — enterprise
    control plane.
  version: 2.0.0
paths:
...
  /api/demo/proof:
    post:
      tags:
      - Demo
      summary: Generate Party Proof
      description: "Generate a deterministic proof for a party record.\n\nSteps:\n\
        \  1. Normalize fields (trim, casing, amount format, timestamp format)\n \
        \ 2. Canonicalize to a deterministic JSON string (sorted keys)\n  3. SHA-256\
        \ hash the canonical JSON\n\nTwo parties applying these same steps to logically-identical\
        \ records\nwill produce identical `proof_hash` values — enabling proof-based\n\
        consistency comparison without exposing raw data."
      operationId: generate_party_proof_api_demo_proof_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyRecord'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/demo/issue:
    post:
      tags:
      - Demo
      summary: Issue Proof
      description: 'Issue a compliance-aware proof for a transaction and persist it
        so it can

        be verified later by an auditor using only the Proof ID.'
      operationId: issue_proof_api_demo_issue_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/demo/verify/{proof_id}:
    get:
      tags:
      - Demo
      summary: Verify Proof By Id
      description: 'Look up a stored proof and re-verify its integrity by re-canonicalizing

        and re-hashing the persisted payload. Returns the compliance state that

        was embedded when the proof was issued.'
      operationId: verify_proof_by_id_api_demo_verify__proof_id__get
      parameters:
      - name: proof_id
        in: path
        required: true
        schema:
          type: string
          title: Proof Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyByIdResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/demo/artifact:
    post:
      tags:
      - Demo
      summary: Build Downloadable Artifact
      description: "Build and sign a standalone proof artifact that can be downloaded,\n\
        shared, and independently verified anywhere.\n\nResponse:\n  - Body: canonical\
        \ JSON of the signed artifact\n  - Content-Type: application/pfp-proof+json;v=1\n\
        \  - Content-Disposition: attachment (suggested filename with proof_id)"
      operationId: build_downloadable_artifact_api_demo_artifact_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/demo/artifact/verify:
    post:
      tags:
      - Demo
      summary: Verify Downloadable Artifact
      description: 'Independently verify an uploaded/pasted signed proof artifact.


        Runs strict schema checks, canonical-ordering enforcement, normalization,

        timestamp skew, constant-time proof_id comparison, kid lookup (including

        revocation), and Ed25519 signature verification.'
      operationId: verify_downloadable_artifact_api_demo_artifact_verify_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Artifact
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactVerifyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ArtifactRequest:
      properties:
        transaction_id:
          type: string
          maxLength: 200
          minLength: 1
          title: Transaction Id
        user_id:
          type: string
          maxLength: 200
          minLength: 1
          title: User Id
        amount:
          anyOf:
          - type: string
          - type: number
          - type: integer
          title: Amount
        compliance:
          $ref: '#/components/schemas/ComplianceResult'
        timestamp:
          anyOf:
          - type: string
          - type: 'null'
          title: Timestamp
      type: object
      required:
      - transaction_id
      - user_id
      - amount
      - compliance
      title: ArtifactRequest
    ArtifactVerifyResult:
      properties:
        valid:
          type: boolean
          title: Valid
        status:
          type: string
          title: Status
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        extracted:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Extracted
      type: object
      required:
      - valid
      - status
      title: ArtifactVerifyResult
...
    ComplianceResult:
      properties:
        kyc:
          type: string
          pattern: ^(Pass|Fail)$
          title: Kyc
        aml:
          type: string
          pattern: ^(Pass|Fail)$
          title: Aml
        limits:
          type: string
          pattern: ^(Within allowed range|Exceeded)$
          title: Limits
        status:
          type: string
          pattern: ^(COMPLIANT|NON-COMPLIANT)$
          title: Status
      type: object
      required:
      - kyc
      - aml
      - limits
      - status
      title: ComplianceResult
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IndustryCheck:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        desc:
          type: string
          maxLength: 500
          minLength: 1
          title: Desc
        status:
          type: string
          pattern: ^(Pass|Fail)$
          title: Status
      type: object
      required:
      - name
      - desc
      - status
      title: IndustryCheck
      description: A single industry-specific compliance check executed before proof
        issuance.
    IndustryContext:
      properties:
        id:
          type: string
          maxLength: 50
          minLength: 1
          title: Id
        label:
          type: string
          maxLength: 100
          minLength: 1
          title: Label
        checks:
          items:
            $ref: '#/components/schemas/IndustryCheck'
          type: array
          maxItems: 20
          minItems: 1
          title: Checks
      type: object
      required:
      - id
      - label
      - checks
      title: IndustryContext
      description: 'Optional industry-aware metadata describing which checks were
        executed.


        When present, this object is embedded in the canonical payload so the

        exact list of industry-specific checks (and their outcomes) is

        cryptographically proven by the issued proof_id. The cryptographic

        primitives (canonicalize → SHA-256 → Ed25519) are unchanged; only the

        input dict gains an additional, well-defined field.'
    IssueRequest:
      properties:
        transaction_id:
          type: string
          maxLength: 200
          minLength: 1
          title: Transaction Id
        user_id:
          type: string
          maxLength: 200
          minLength: 1
          title: User Id
        amount:
          anyOf:
          - type: string
          - type: number
          - type: integer
          title: Amount
        created_at:
          type: string
          minLength: 1
          title: Created At
        compliance:
          $ref: '#/components/schemas/ComplianceResult'
        industry:
          anyOf:
          - $ref: '#/components/schemas/IndustryContext'
          - type: 'null'
      type: object
      required:
      - transaction_id
      - user_id
      - amount
      - created_at
      - compliance
      title: IssueRequest
    IssueResponse:
      properties:
        proof_id:
          type: string
          title: Proof Id
        proof_hash:
          type: string
          title: Proof Hash
        transaction_id:
          type: string
          title: Transaction Id
        issued_at:
          type: string
          title: Issued At
        metadata:
          $ref: '#/components/schemas/ProofMetadata'
      type: object
      required:
      - proof_id
      - proof_hash
      - transaction_id
      - issued_at
      - metadata
      title: IssueResponse
...
    PartyRecord:
      properties:
        transaction_id:
          type: string
          maxLength: 200
          minLength: 1
          title: Transaction Id
        user_id:
          type: string
          maxLength: 200
          minLength: 1
          title: User Id
        amount:
          anyOf:
          - type: string
          - type: number
          - type: integer
          title: Amount
        created_at:
          type: string
          minLength: 1
          title: Created At
      type: object
      required:
      - transaction_id
      - user_id
      - amount
      - created_at
      title: PartyRecord
    ProofMetadata:
      properties:
        algorithm:
          type: string
          title: Algorithm
        canonicalization:
          type: string
          title: Canonicalization
        generated_at:
          type: string
          title: Generated At
      type: object
      required:
      - algorithm
      - canonicalization
      - generated_at
      title: ProofMetadata
    ProofResponse:
      properties:
        proof_hash:
          type: string
          title: Proof Hash
        normalized_payload:
          additionalProperties: true
          type: object
          title: Normalized Payload
        metadata:
          $ref: '#/components/schemas/ProofMetadata'
      type: object
      required:
      - proof_hash
      - normalized_payload
      - metadata
      title: ProofResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerifyByIdResponse:
      properties:
        valid:
          type: boolean
          title: Valid
        proof_id:
          type: string
          title: Proof Id
        transaction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Transaction Id
        compliance:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Compliance
        industry:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Industry
        issued_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Issued At
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - valid
      - proof_id
      title: VerifyByIdResponse
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://pfp-evidence.preview.emergentagent.com
  description: Sandbox/preview
- url: https://api.pfprotocol.com
  description: Production